Add the expand method.
The expand method was added to the following algorithms:

- `WS_NC_MULT`
- `B_WS_NC_MULT`

Where the `B_WS_NC_MULT` `expand` method has the property of update
two arrays in each call. This could be a issue in the performance to
the future. Let's see how this works and we will update according to
ours observations.